Move Mux out of array.py
